IGNITE-26358 Many fixes

+++ modules/platforms/python/cpp_module/py_connection.h
@@ -41,7 +41,7 @@
  * @return A new connection class instance.
  */
 PyObject* make_py_connection(std::vector<ignite::end_point> addresses, const char* schema, const char* identity,
-    const char* secret, int page_size, int timeout, float heartbeat_interval, bool autocommit, ssl_config &&ssl_cfg);
+    const char* secret, int page_size, int timeout, double heartbeat_interval, bool autocommit, ssl_config &&ssl_cfg);
 
 /**
  * Prepare PyConnection type for registration.
